справочники, изменения в файлах апи
