+++ README.md
@@ -5,6 +5,8 @@ HRSpace - маркетплейс частных рекрутеров и кадр
 
 Проект состоит из бэкенд-приложения на Django и фронтенд-приложения на React.
 
+[Схема базы данных](https://dbdiagram.io/d/HRspase-65ef0862b1f3d4062ca0f7fc) 
+
 ## Возможности проета:
 
 ## Запуск:
